chore: move container to containers folder
